♻️ Refactor: Rename app to redux, api to axios
