+++ client/src/features/library/components/LibraryHeader/HeaderTitle.jsx
@@ -2,7 +2,7 @@ import styles from './HeaderTitle.module.scss';
 import { RiAddLargeFill, RiArchiveStackLine } from 'react-icons/ri';
 import IconButton from '../../../../ui/IconButton/IconButton.tsx';
 import { createPlaylist } from '../../../playlist/playlistThunks.ts';
-import { useAppDispatch } from '../../../../app/hooks.ts';
+import { useAppDispatch } from '../../../../redux/hooks.ts';
 
 // INFO: Collapse hides "Your Library" text and "+" buttons
 const HeaderTitle = ({ isCollapsed, onCollapse }) => {
