+++ client/src/features/playlist/components/PlaylistActions.tsx
@@ -2,7 +2,7 @@ import PlayHeader from '../../../ui/PlayHeader/PlayHeader.tsx';
 import PlayButton from '../../../ui/PlayHeader/PlayButton.tsx';
 import TransparentButton from '../../../ui/Button/TransparentButton.tsx';
 import { Playlist } from '../playlistSlice.ts';
-import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts';
+import { useAppDispatch, useAppSelector } from '../../../redux/hooks.ts';
 import { removePlaylist, savePlaylist } from '../playlistThunks.ts';
 import PlaylistHeaderActions from './PlaylistHeaderActions.tsx';
