+++ client/src/features/playlist/components/PlaylistHeaderActions.tsx
@@ -4,7 +4,7 @@ import { RiDeleteBin6Line, RiFileCopyLine, RiMoreFill } from 'react-icons/ri';
 import styles from './PlaylistActions.module.scss';
 import DropdownList from '../../../ui/Dropdown/DropdownList.tsx';
 import DropdownItem from '../../../ui/Dropdown/DropdownItem.tsx';
-import { useAppDispatch } from '../../../app/hooks.ts';
+import { useAppDispatch } from '../../../redux/hooks.ts';
 import { deletePlaylist } from '../playlistThunks.ts';
 import { useNavigate } from 'react-router-dom';
 import toast from 'react-hot-toast';
