+++ client/src/features/userPage/components/UserProfile.tsx
@@ -1,6 +1,6 @@
 import GradientBackground from '../../../ui/GradientBackground/GradientBackground.tsx';
 import LoadingScreen from '../../../ui/StatusScreens/LoadingScreen.tsx';
-import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts';
+import { useAppDispatch, useAppSelector } from '../../../redux/hooks.ts';
 import { followUser, unfollowUser } from '../../auth/userThunks.ts';
 import ImageHeader from '../../../ui/ImageHeader/ImageHeader.tsx';
 import { useNavigate, useParams } from 'react-router-dom';
